initial commit
This entire directory and repository was created when i wanted
to start writing a p2p chat app in python. It just started bogging
my brain and i decided fuck it lets do this! I really do not know
much about p2p systems or distributed for that matter. I started
exploring that field earlier this year and i have fallen in love
with the theory  but not had much experience practicing it. So here
goes me struggling to do write a fully distibuted p2p chat service in
golang and python.