alot of changes at this point
Im trying to stabilise the thing because everyday, i have new ideas
and i keep on adding them to the packages and its starting to piss me
off. I will rewrite the ipaddress verification code and use the
ipaddr package. What i have done so far is just trying to redesign
the whole api of the package to be just like the golang net package
and make it possible for me to add new layers and shit like that.