document functions,methods and classes
Most of the things in the package are not properly documented.
I think they are now. I redid the typing for most of the functions
and classes and rewrote most of the comment docs. The api of the
package is almost similar for all types now. It should work but
i've not tested it soo. I'll come back and test it later. I'm not
writing tests thoo. Im just testing by using the package to see if it
works. Next up after testing is making sure the ip address parsing
and resolving also has a better interface and works as expected which
is the way golang net.Addr works.