think i'm done with address resolution
this was a tough one, not completely done but more will come from
using the package. This is created to be extended to be used with
any socket and address type that any os supports. Most or all of the
logic in this packages are either copied from golang implementation
of their net package or inspired by it. With that being said, the
golang net implementation is a nice piece of work. Nice work folks
of the golang team.